==> notebooks/01-data-exploration.ipynb.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

#load the data
train = pd.read_csv('../data/raw/train.csv')

#explanitory analysis (EDA)

train.info()

print(train.head())

train.shape

#Display unique counts per column with formatted output
train.nunique().to_frame().rename(columns={0: 'Unique Values'}).style.format({'Unique Values': '{:,.0f}'})

# looking at the response variable. The data is positivly sqewed. Maybe there are some outliers. 
plt.hist(train['SalePrice'])
plt.show()
train['SalePrice'].min()

#explore the correlation of each vairable to the sale price

# Identify numerical columns
numerical_columns = train.select_dtypes(include=["int64", "float64"]).columns

# Plot distribution of each numerical feature
plt.figure(figsize=(14, len(numerical_columns) * 3))
for idx, feature in enumerate(numerical_columns, 1):
    plt.subplot(len(numerical_columns), 2, idx)
    plt.scatter( train[feature],train['SalePrice'])
    plt.title(f"{feature}")

# Adjust layout and show plots
plt.tight_layout()
plt.show()

#Thigs that Positivily affect the sale price: 
#TotalBsmtSF, 1stFlrSF, 2ndFlrSF, GrLivArea, FullBath, TotRmsAbvGrd, GarageYrBlt, GarageArea

#Things that have no effect (or too few information)- consider dropping these features. 
#MiscVal, PoolArea

#exploring the sqewness of the vairables

# Plot distribution of each numerical feature
plt.figure(figsize=(14, len(numerical_columns) * 3))
for idx, feature in enumerate(numerical_columns, 1):
    plt.subplot(len(numerical_columns), 2, idx)
    sns.histplot(train[feature], kde=True)
    plt.title(f"{feature} | Skewness: {round(train[feature].skew(), 2)}")

# Adjust layout and show plots
plt.tight_layout()
plt.show()

#explore non numerical features:
# Identify non-numberical columns
non_numerical_columns = train.select_dtypes(include=['object']).columns

# Plot distribution of each numerical feature
plt.figure(figsize=(14, len(non_numerical_columns) * 3))
for idx, feature in enumerate(non_numerical_columns, 1):
    plt.subplot(len(non_numerical_columns), 2, idx)
    sns.boxplot(x=train[feature], y=train['SalePrice'])
    plt.title(f"{feature}")

# Adjust layout and show plots
plt.tight_layout()
plt.show()

#NA has been left out. Can we find a way to keep it in?

#Zone RL is most expensive, Street Pave is most expensive, paved alleys are more expensive, 2 story, 
#exc qualatiy material of external,basement, heating and air con, good kitchen, fireplace,  

#Explore category cardinality. Most are low so one hot encoding is ok.
non_numerical_columns = train.select_dtypes(include=['object']).nunique()
non_numerical_columns[non_numerical_columns > 10]
