==> notebooks/02-preprocessing.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split 

#load the data
train = pd.read_csv('../data/raw/train.csv')

#na cleanup

#basements
basements=train[['BsmtQual','BsmtCond','BsmtExposure','BsmtFinType1','BsmtFinType2']]
basements[basements.isnull().any(axis=1)]
#index 948 and 332 are the problem. drop those rows

#electrical
electrical=train['Electrical']
electrical[electrical.isnull()]

#replace LotFrontage with median of that column
train['LotFrontage'].median()
train['LotFrontage']=train['LotFrontage'].fillna(69).astype(int)

#drop all rows with problems
train_processed = train.drop([train.index[948], train.index[332],train.index[1379]])

#change all nas to 'none'
train_processed.fillna('None', inplace=True)

#normalize dependant variable to minimize outlier effect
train_processed['SalePrice'] = np.log(train_processed['SalePrice'])

#split data into train test
train_data,test_data = train_test_split(train_processed,random_state=104,test_size=0.2,shuffle=True) 

# Save data to 'data/processed/' folder
train_data.to_csv('../data/processed/train_data.csv', index=False)
test_data.to_csv('../data/processed/test_data.csv', index=False)
